requirements.txt
openai=1.98.0
edge-tts>=1.1.1
faster-whisper>=0.4.3
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
requests>=2.28.0
sounddevice>=0.4.6
scipy>=1.11.0
